Use rank model from Clinical Genomics Stockholm
The problem with doing this is that it doesn't work. Their annotation
process is different from ours, resulting in field names not matching,
or are missing completely. Still, this can serve as an example of how to
handle the versioning of the rank model, but changes are needed. We can
either conform to their annotation workflow, or we create our own
repository of rank model configs that are adapted to our current
workflow.