feat: add VCF filtering
This is supposed to be a general filtering rule for VCF files. There are
still some things to be addressed before this is actually the case
though.

One thing that bothers me a bit with this filtering approach is that it
only enables hard filtering. I'm using vembrane for applying the
filters, and although I really like the application, there is no option
for soft filtering, at least that I have seen.