fix: correct header order for vembrane
vcfanno reordered the header so that the software line for VEP came
after the CSQ INFO line. This in turn resulted in vembrane crashing
since it was depending on the opposite order. To solve this, I swapped
the execution order for vcfanno and most_severe_csq and did the
reordering in most_severe_consequence.py.