fix: allow missing panel and filtering columns
The default in these cases should be to not apply any filtering. There
was a discrepancy where the column had to be defined, but values in the
column could be missing. With this change it should work both if the
columns themselves are missing, or if individual samples have missing
values in these columns.