fix: respect the Number set in the header for AF/AO
Even though the number for these two should be "A", I have seen cases
where they are "1", and also cases where one of them is "A" and the
other is "1". This doesn't really make sense, but the pysam library
doesn't allow me to modify the header the way I want to (for good
reason).

Tests were added, and everything looks reasonable.