Integrate default config in package
This way the default config is applied automatically when running, and
supplying a config file on the command line will overwrite the defaults.
This way any config supplied by a user can be minimal. The defaults will
have to be clearly documented though, and at the moment, many of these
paths are specific to our development server.

Changes that should be done are:

- Clear all path config parameters.
- Make a CLI command for generating a template config that can be filled
  out by the user.

Another issue with this is how to deal with relative paths, e.g. other
config files that are included in the package. Right now, this will
break things if they are not overwritten.