feat: wrap pipeline in a proper package (#47)
* Start conversion to a proper python package

For this I'm using poetry.

* Fix the AF fixing test and add corner cases

The script is currently a mess with a lot of branching. This should be
simplified.

* Refactor cli into modules

* Move panels into package

* Ignore build directory

* Fix paths in integration test

* Update README to reflect changes

* Utility function for getting snakefile path

* Integrate default config in package

This way the default config is applied automatically when running, and
supplying a config file on the command line will overwrite the defaults.
This way any config supplied by a user can be minimal. The defaults will
have to be clearly documented though, and at the moment, many of these
paths are specific to our development server.

Changes that should be done are:

- Clear all path config parameters.
- Make a CLI command for generating a template config that can be filled
  out by the user.

Another issue with this is how to deal with relative paths, e.g. other
config files that are included in the package. Right now, this will
break things if they are not overwritten.

* Add CLI integration tests

* Proper return code from CLI commands

* Evaluate filter config files relative to main config

* Make relative path check slightly more general

* Minor clarification

ci: add CLI integration tests
feat: ability to set apptainer arguments from the CLI