feat: trio annotation functionality (#49)
* Add capability to annotate trios

This required quite a bit of refactoring since now families are the
focal point, and not samples.

* Add rule graph to readme

* Detail how to skip integration tests

* Better logging

* Fix illegal family IDs

* Make sure integration test directries are added

* Add trio CLI command

* Minor rule graph clarification

* Update usage

* Mark more output files as temp

* Remove temp for filtered VCF

For some reason, setting this one as temp breaks the pipeline. I'm
guessing that snakemake fails to detect that it is needed further
downstream and then deletes it. This probably has a different solution,
but this fixes the immediate problem.

* Do not create multiallelic variants when merging

Scout does not like multiallelic variants, so these have to be separate
entries in the final VCF.

* Add peddy

* Fix integration tests for trios with peddy

* Update rank model variable names for consistency

* Add peddy files to load config

* Flag peddy output as temp

* Remove commented code

* Make sure samples and VCFs are paired correctly

* Put the correct peddy files/paths in the load config

doc: add rule graph to README