Run only one single-sample CLI test
They were both more or less doing the same thing, and running the only
one in an isolated environment minimises the cleanup that has to be done
now and then. The downside is that the tests always would have to be run
all the way from start to finish, but I'm willing to live with that.