fix: allow for numbers at the beginning of a sample name
I don't remember the rationale for not allowing numbers at the beginning
of a sample name, but I think that is overly restrictive, especially
considering that some of our old IDs do start with numbers. This allows
us to use those IDs all the way through, and that should minimise any
confusion.