fix: address issue with empty string gene panels
When running the pipeline with no gene panel set, the schema validation
converted this to an empty string. This in turn triggered a lot of extra
rules for no reason when the panel filtering wasn't required. For some
reason this didn't crash downstream.

This commit addresses this by setting the default to `null` in the JSON
schema. Test cases were added to the integration tests that will catch
this issue. It would have been nicer to have this as a unit test, but
this requires a refactor of the sample handling, and this is something I
will get to at a later time.